refactor(searchcore): move queue into module + Queue interface; 全量改用, 删 queue (P1.5)
