+++ internal/core/documents/storage_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/codetrek/haystack/internal/conf"
 	"github.com/codetrek/haystack/internal/core/storage"
-	"github.com/codetrek/haystack/internal/utils/queue"
+	"github.com/codetrek/haystack/searchcore/queue"
 	"github.com/stretchr/testify/assert"
 )
